Fixed authorization

+++ individueel-project-1.3-api/Dto/RoomRequestDto.cs
@@ -4,8 +4,8 @@ public class RoomRequestDto
 {
     public required Guid RoomId { get; init; }
     public required string Name { get; init; }
-    public required float Width { get; init; }
-    public required float Height { get; init; }
+    public required int Width { get; init; }
+    public required int Height { get; init; }
     public required string TileId { get; init; }
     public bool? IsOwner { get; init; }
     public List<PropRequestDto> Props { get; init; } = [];
