+++ individueel-project-1.3-api/Models/Room.cs
@@ -8,8 +8,8 @@ public class Room
 
     public required Guid RoomId { get; init; }
     public required string Name { get; init; }
-    public required float Width { get; init; }
-    public required float Height { get; init; }
+    public required int Width { get; init; }
+    public required int Height { get; init; }
     public required string TileId { get; init; }
 
     public List<UserEntry> Users { get; set; } = [];
